Diving - WritePost => Container & Presenter
